Move add-existing-organization-dialog.component to AC
